fixing sidebar in example
