--- src/styles/pages/example.module.css
@@ -1,34 +1,0 @@
-.root {
-  & p {
-    margin-top: var(--gutter-half);
-  }
-}
-
-.collapsed {
-  margin-left: 0;
-  transition: 0.4s ease;
-}
-
-.expanded {
-  margin-left: 100px;
-  transition: 0.4s ease;
-}
-
-.section {
-  margin-bottom: var(--vertical-margin-large);
-}
-
-.placeholder {
-  background-color: var(--gray);
-  height: 150px;
-  width: 150px;
-  padding-right: var(--gutter);
-}
-
-.cover {
-  margin-bottom: var(--gutter);
-  margin-top: var(--gutter);
-  & img {
-    width: 100%;
-  }
-}
