+++ src/templates/reference/function.js
@@ -13,13 +13,12 @@ import { CodeList, ExampleList } from '../../components/ReferenceItemList';
 import { ExampleItem } from '../../components/ExamplesList';
 import { widont } from '../../utils/index.js';
 
-import { useTree, useHighlight } from '../../hooks';
+import { useTree, useHighlight, useSidebar } from '../../hooks';
 import {
   usePreparedItems,
   usePreparedExamples,
   usePreparedList,
-  useInUseExamples,
-  useSidebar
+  useInUseExamples
 } from '../../hooks/reference';
 import { referencePath } from '../../utils/paths';
 
@@ -78,7 +77,9 @@ const RefTemplate = ({ data, pageContext, ...props }) => {
               </Section>
             )}
             <Section title={intl.formatMessage({ id: 'description' })}>
-              <p dangerouslySetInnerHTML={{ __html: widont(entry.description) }} />
+              <p
+                dangerouslySetInnerHTML={{ __html: widont(entry.description) }}
+              />
             </Section>
             {examples && (
               <Section
